Fixes auto-config of loggers
In a previous revision I attempted to set things up so a CSV logger was
always automatically configured. This does not play nice with W&B sweeps
though: for reasons I don't fully undersatnd, it would:

1. rename every run in a sweep to "lightning_logs" after initialization
2. attempt and fail to write `config.yaml` to the same location, causing
   a failure on later runs.

I just disable this and document it.

Some minor drive-bys are done on sample config files as well.